chore: migrate password modals (#1371)
* chore: WIP moving modals to WalletProvider

* fix: native password and kk pin modal migrated

* fix: migrated the keepkey passphrase modal to live under wallet context
