+++ src/context/WalletProvider/NativeWallet/hooks/useNativeEventHandler.ts
@@ -1,28 +1,27 @@
 import { Event } from '@shapeshiftoss/hdwallet-core'
 import { NativeEvents } from '@shapeshiftoss/hdwallet-native'
 import { Dispatch, useEffect } from 'react'
-import { ActionTypes } from 'context/WalletProvider/actions'
+import { ActionTypes, WalletActions } from 'context/WalletProvider/actions'
 import { InitialState } from 'context/WalletProvider/WalletProvider'
-import { useModal } from 'hooks/useModal/useModal'
 
-type KeyringState = Pick<InitialState, 'keyring' | 'walletInfo'>
+type KeyringState = Pick<InitialState, 'keyring' | 'walletInfo' | 'modal'>
 
 export const useNativeEventHandler = (state: KeyringState, dispatch: Dispatch<ActionTypes>) => {
-  const { nativePassword } = useModal()
-  const { keyring } = state
+  const { keyring, modal } = state
 
   useEffect(() => {
     const handleEvent = (e: [deviceId: string, message: Event]) => {
       console.info('Native Wallet Event', e)
+      const deviceId = e[0]
       switch (e[1].message_type) {
         case NativeEvents.MNEMONIC_REQUIRED:
-          if (!nativePassword.isOpen) {
-            nativePassword.open({ deviceId: e[0] })
-          }
+          if (!deviceId) break
+          dispatch({ type: WalletActions.NATIVE_PASSWORD_OPEN, payload: { modal: true, deviceId } })
+
           break
         case NativeEvents.READY:
-          if (!nativePassword.isOpen) {
-            nativePassword.close()
+          if (modal) {
+            dispatch({ type: WalletActions.SET_WALLET_MODAL, payload: false })
           }
           break
         default:
@@ -31,6 +30,7 @@ export const useNativeEventHandler = (state: KeyringState, dispatch: Dispatch<Ac
       }
     }
 
+    // if (!nativePassword) return
     if (keyring) {
       keyring.on(['Native', '*', NativeEvents.MNEMONIC_REQUIRED], handleEvent)
       keyring.on(['Native', '*', NativeEvents.READY], handleEvent)
@@ -39,5 +39,5 @@ export const useNativeEventHandler = (state: KeyringState, dispatch: Dispatch<Ac
       keyring.off(['Native', '*', NativeEvents.MNEMONIC_REQUIRED], handleEvent)
       keyring.off(['Native', '*', NativeEvents.READY], handleEvent)
     }
-  }, [dispatch, nativePassword, keyring])
+  }, [dispatch, keyring, modal])
 }
