+++ src/context/WalletProvider/config.ts
@@ -2,6 +2,8 @@ import { ComponentWithAs, IconProps } from '@chakra-ui/react'
 import { RouteProps } from 'react-router-dom'
 
 import { KeepKeyConnect } from './KeepKey/components/Connect'
+import { KeepKeyPassphrase } from './KeepKey/components/Passphrase'
+import { KeepKeyPin } from './KeepKey/components/Pin'
 import { KeepKeySuccess } from './KeepKey/components/Success'
 import { KeepKeyConfig } from './KeepKey/config'
 import { KeyManager } from './KeyManager'
@@ -9,6 +11,7 @@ import { MetaMaskConnect } from './MetaMask/components/Connect'
 import { MetaMaskFailure } from './MetaMask/components/Failure'
 import { MetaMaskSuccess } from './MetaMask/components/Success'
 import { MetaMaskConfig } from './MetaMask/config'
+import { EnterPassword } from './NativeWallet/components/EnterPassword'
 import { NativeCreate } from './NativeWallet/components/NativeCreate'
 import { NativeImport } from './NativeWallet/components/NativeImport'
 import { NativeLoad } from './NativeWallet/components/NativeLoad'
@@ -41,14 +44,17 @@ export const SUPPORTED_WALLETS: Record<KeyManager, SupportedWalletInfo> = {
       { path: '/native/import', component: NativeImport },
       { path: '/native/create', component: NativeCreate },
       { path: '/native/create-test', component: NativeTestPhrase },
-      { path: '/native/success', component: NativeSuccess }
+      { path: '/native/success', component: NativeSuccess },
+      { path: '/native/enter-password', component: EnterPassword }
     ]
   },
   [KeyManager.KeepKey]: {
     ...KeepKeyConfig,
     routes: [
       { path: '/keepkey/connect', component: KeepKeyConnect },
-      { path: '/keepkey/success', component: KeepKeySuccess }
+      { path: '/keepkey/success', component: KeepKeySuccess },
+      { path: '/keepkey/enter-pin', component: KeepKeyPin },
+      { path: '/keepkey/passphrase', component: KeepKeyPassphrase }
     ]
   },
   [KeyManager.MetaMask]: {
